[Decomp][Player] Cohesion split: extract combo-point methods to PlayerCombo.cpp
